Add directfs and rdbplugin, update fs and dir
Homing in on a first cut of the control protocol. I found myself adding
the labels and status maps for most objects, so I moved them out into
common.proto. We're not optimising for performance here.

This is still a shell. I think I've encapsulated most current configfs
functionality in a way that makes sense.

Where the client programs (reaper, symmetra) need configuration it makes
sense for them to be able to connect to their 'backend' programs
(symmetra talks to fs, reaper talks - I think - to the rdpplugin) to do
their work.

I've not added services for symmetra and reaper because it's hard to see
how we'd actually communicate with those programs the way they're
currently spawned - to what would we connect? They will have to connect
somewhere to fetch configuration, though, otherwise they don't know
volume sizes. That sounds like a bigger problem than it is - symmetra
already calls into FS to do its sync, for example.

Actually I believe the ultimate solution there is to implement both
reaper and symmetra as gRPC calls into the plugin and the FS
respectively, as that's easy to do and keeps the functionality
localised. A conversation for another day.