Fixed message copying
Got this whole 'copying message' thing figured out. Previously attempts
to copy messages were not working properly, with exceptions thrown by
MergeFrom().

After all sorts of experimentation, it turns out that the problems only
exist for fields of type Message, i.e. a 'submessage'. For scalar fields
it's easy to just copy those directly by setting them in the constructor
for the new message object.

For message fields, check it's present in the original message
(HasField()) and copy if it's present using CopyFrom(). That's been
wrapped in mock_common.copy_submessage() to keep it tidy.

Now, each time we're copying messages around, just have a copy_*()
function that knows about the message types. It'd be nicer to have this
done automatically, but this works. It's not as if the real
implementation is going to be in Python...