Rename generation->revision, add flag
Use 'revision' instead of 'generation' to be consistent with etcd's
terminology. As req'd by karolis@.

Add a flag indicating that versioning is active for this object. Without
this flag being true, version information will be ignored. This allows
for a gradual enablement of versioning, and makes command line tools
and tests less of a pain in the arse to use.