make: fix generating mocks
Previously mocks were not updated when using the "go" makefile target.

When running the go-mocks makefile target, it would use the $GOPATH
files, not the local files, generating incorrect mocks.

Adds mockgen to the dapper container.

Updates the Go version, because once gomock was added to the dapper
build, it failed due to the old version.

Moves the folder to $GOPATH when generating the mocks so it works in
dapper / generates the correct code.