Add UUID string field everywhere we take an inode
Mostly for query messages, but in some other cases too, add a way to
specify UUIDs (strings) instead of inodes. For the most part the query
APIs are unused at the moment, which buys us some time.

Stats will need revisited anyway, but at the moment it's all indexed by
inode which is going to cause us problems. I can't fix that now.

The most important addition, though, is that we have a UUID target field
for FsPresentation and DirectorPresentation. Without that, we have no
way of linking a presentation to its target. We can't use the
DataplaneCommon hack we're using elsewhere because the target UUID isn't
anywhere to be found.