Mirror the presentation filename
A longstanding modelling error put the filename for the presentation
object on the target volume object. That was done because Karolis
explained that it was modelled that way in the control plane, but it was
a mistake by me because it makes spaghetti out of our FUSE filesystem
code, where we have to look up the filename for presentation objects via
their volumes, requiring indexes and much complexity.

It also feels super-racy to depend on these objects being perfectly in
sync. I have no evidence to back that up, it's a gut feel thing.

In any case, the string filename field on FsPresentation should for the
moment be the same contents as the presentation_filename field on
FsVolume. Once things have calmed down a bit - hah! - we can deprecate
the FsVolume presentation_filename field and reduce complexity in the
FUSE code to match.