Standalone Docker build image.
Actually a dapper build image, but that's ok.

There's no need to use the gigantic Dataplane build image for this, and
it means that we're decoupling this source-critical chunk of c2 from
dataplane's build environment.

Only install the tools we need, except that we're also installing the
C++ gRPC toolchain so we can generate using buf there too. It means we
can use buf's fancy generation and linting tools in one place.